moving classes to a fdm package
