+++ src/main/scala/cognite/spark/v1/fdm/FlexibleDataModelCorePropertyRelation.scala
@@ -1,17 +1,12 @@
-package cognite.spark.v1
+package cognite.spark.v1.fdm
 
 import cats.Apply
 import cats.effect.IO
 import cats.implicits._
-import cognite.spark.v1.FlexibleDataModelBaseRelation.ProjectedFlexibleDataModelInstance
-import cognite.spark.v1.FlexibleDataModelRelationFactory.ViewCorePropertyConfig
-import cognite.spark.v1.FlexibleDataModelRelationUtils.{
-  createEdgeDeleteData,
-  createEdges,
-  createNodeDeleteData,
-  createNodes,
-  createNodesOrEdges
-}
+import cognite.spark.v1.fdm.FlexibleDataModelBaseRelation.ProjectedFlexibleDataModelInstance
+import cognite.spark.v1.fdm.FlexibleDataModelRelationFactory.ViewCorePropertyConfig
+import cognite.spark.v1.fdm.FlexibleDataModelRelationUtils._
+import cognite.spark.v1.{CdfSparkException, CdfSparkIllegalArgumentException, CdpConnector, RelationConfig}
 import com.cognite.sdk.scala.v1.GenericClient
 import com.cognite.sdk.scala.v1.fdm.common.filters.FilterDefinition
 import com.cognite.sdk.scala.v1.fdm.common.properties.PropertyDefinition.ViewPropertyDefinition
