+++ src/test/scala/cognite/spark/v1/fdm/utils/FlexibleDataModelTestBase.scala
@@ -2,7 +2,8 @@ package cognite.spark.v1.fdm.utils
 
 import cats.effect.IO
 import cats.{Applicative, Apply}
-import cognite.spark.v1.{DefaultSource, FlexibleDataModelRelationFactory, SparkTest}
+import cognite.spark.v1.fdm.FlexibleDataModelRelationFactory
+import cognite.spark.v1.{DefaultSource, SparkTest}
 import com.cognite.sdk.scala.v1.fdm.common.properties.PropertyDefinition.{ContainerPropertyDefinition, ViewCorePropertyDefinition}
 import com.cognite.sdk.scala.v1.fdm.common.properties.PropertyType.DirectNodeRelationProperty
 import com.cognite.sdk.scala.v1.fdm.common.properties.{ListablePropertyType, PrimitivePropType, PropertyType}
